update structure to mvc
