+++ src/view/index.js
@@ -1,7 +1,7 @@
-import Register from './register-view.js';
-import Login from './login-view.js';
+import viewRegister from './register-view.js';
+import viewLogin from './login-view.js';
 
 export const components = {
-  register: Register,
-  login: Login,
+  register: viewRegister,
+  login: viewLogin,
 };
